# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Argus Alpha API
  version: 1.0.0-alpha
  description: Contract-first Alpha API for chat-first investing idea validation.
servers:
  - url: /api/v1
paths:
...
  /api/v1/auth/guest:
    post:
      description: Create or reuse one verified anonymous guest session when server policy permits it.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GuestBootstrapRequest"
      responses:
        "200":
          description: Guest session and temporary workspace are ready.
        "403":
          description: Guest access is disabled or the browser origin is rejected.
        "429":
          description: Anonymous-session creation is temporarily rate limited.
        "503":
          description: Anonymous Auth or temporary-workspace bootstrap is unavailable.
...
  /api/v1/auth/logout:
    post:
      responses:
        "200":
          description: Session cleared.
        "403":
          description: Untrusted browser origin rejected.
  /api/v1/auth/session:
    get:
      responses:
        "200":
          description: Current authentication state and user.
        "503":
          $ref: "#/components/responses/AuthSessionVerificationUnavailable"
...
components:
  responses:
    AuthSessionVerificationUnavailable:
      description: Bounded Supabase session verification is temporarily unavailable; retry without redirecting to login.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            type: https://api.argus.app/problems/auth-session-verification-unavailable
            title: Session Verification Unavailable
            status: 503
            detail: Argus could not verify this session. Please try again.
            code: auth_session_verification_unavailable
            request_id: req_123
  schemas:
    GuestBootstrapRequest:
      type: object
      required:
        - captcha_token
      properties:
        captcha_token:
          type: string
          minLength: 1
          maxLength: 4096
        language:
          type: string
          enum: [en, es-419]
          default: en
...
    Error:
      type: object
      description: RFC 9457 Problem Details with code and request_id.
...